FPRL-392: Going to court e2e tests, updated
