+++ test/end-to-end/tests/GoingToCourt.js
@@ -6,8 +6,8 @@ Scenario('Going to court event - basic journey', I => {
   I.goingToCourtSelectNoForAll();
 }).retry({ retries: 3, minTimeout: 30000 });
 
-Scenario('Going to court event - complex journey', I => {
-  I.loginAsSolicitor();
-  I.createCase();
-  I.goingToCourtSelectYesForAll();
-}).retry({ retries: 3, minTimeout: 30000 });
+// Scenario('Going to court event - complex journey', I => {
+//   I.loginAsSolicitor();
+//   I.createCase();
+//   I.goingToCourtSelectYesForAll();
+// }).retry({ retries: 3, minTimeout: 30000 });
