+++ test/end-to-end/tests/safeguardingAndRiskOfHarm_test.js
@@ -4,4 +4,4 @@ Scenario('Create an event and Submit the Safeguarding and Risk of Harm workflow'
   await I.loginAsSolicitor();
   await I.createCase();
   await I.safeguardingAndRiskOfHarm();
-}).retry({retries: 3, minTimeout: 30000});
+}).retry({ retries: 3, minTimeout: 30000 });
